requirements.txt
tox==2.6.0
six==1.12.0
channels==2.3.1
codeclimate-test-reporter
pytest-asyncio==0.10.0
pytest-django==3.6.0